+++ mlir/lib/Dialect/LLVMIR/Transforms/CMakeLists.txt
@@ -7,7 +7,6 @@ add_mlir_dialect_library(MLIRLLVMIRTransforms
   LegalizeForExport.cpp
   OptimizeForNVVM.cpp
   RequestCWrappers.cpp
-  OpenMPOffloadPrivatizationPrepare.cpp
 
   DEPENDS
   MLIRLLVMPassIncGen
